Added TextDegree Filter
